# docs/openapi.yaml
basePath: /api/v1
definitions:
...
  github_com_uigraph_app_internal_httputil.errorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: REST API for uigraph.
  title: uigraph API
  version: "1.0"
paths:
...
  /orgs/{orgID}/maps:
    get:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: ListMaps
      tags:
      - maps
    post:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: CreateMap
      tags:
      - maps
  /orgs/{orgID}/maps/{mapID}:
    delete:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: mapID
        in: path
        name: mapID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: DeleteMap
      tags:
      - maps
    get:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: mapID
        in: path
        name: mapID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: GetMap
      tags:
      - maps
    put:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: mapID
        in: path
        name: mapID
        required: true
        type: string
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: UpdateMap
      tags:
      - maps
...
  /orgs/{orgID}/maps/{mapID}/frames/{frameID}/links:
    get:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: mapID
        in: path
        name: mapID
        required: true
        type: string
      - description: frameID
        in: path
        name: frameID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: ListLinks
      tags:
      - maps
    post:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: mapID
        in: path
        name: mapID
        required: true
        type: string
      - description: frameID
        in: path
        name: frameID
        required: true
        type: string
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: CreateLink
      tags:
      - maps
  /orgs/{orgID}/maps/{mapID}/frames/{frameID}/links/{linkID}:
    delete:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: mapID
        in: path
        name: mapID
        required: true
        type: string
      - description: frameID
        in: path
        name: frameID
        required: true
        type: string
      - description: linkID
        in: path
        name: linkID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: DeleteLink
      tags:
      - maps
    put:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: mapID
        in: path
        name: mapID
        required: true
        type: string
      - description: frameID
        in: path
        name: frameID
        required: true
        type: string
      - description: linkID
        in: path
        name: linkID
        required: true
        type: string
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: UpdateLink
      tags:
      - maps
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"